# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FranchiseIntelligenceOS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /compliance:
    get:
      operationId: listComplianceChecks
      tags: [compliance]
      summary: List compliance checks (optionally by location)
      parameters:
        - name: locationId
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: List of compliance checks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComplianceCheck"
    post:
      operationId: createComplianceCheck
      tags: [compliance]
      summary: Create a compliance check
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplianceCheckInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceCheck"
  /compliance/{id}:
    patch:
      operationId: updateComplianceCheck
      tags: [compliance]
      summary: Update a compliance check
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplianceCheckUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceCheck"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
    delete:
      operationId: deleteComplianceCheck
      tags: [compliance]
      summary: Delete a compliance check
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /audits:
    get:
      operationId: listAudits
      tags: [audits]
      summary: List audits (optionally by location)
      parameters:
        - name: locationId
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: List of audits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Audit"
    post:
      operationId: createAudit
      tags: [audits]
      summary: Create an audit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuditInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Audit"
  /audits/{id}:
    delete:
      operationId: deleteAudit
      tags: [audits]
      summary: Delete an audit
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents (optionally by location or task)
      parameters:
        - name: locationId
          in: query
          required: false
          schema: { type: integer }
        - name: taskId
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a document (SOP, policy, or note; content is searchable by the AI assistant)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document with content
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /ai/conversations/{id}/messages:
    get:
      operationId: listConversationMessages
      tags: [ai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
  /ai/conversations/{id}:
    delete:
      operationId: deleteConversation
      tags: [ai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /ai/chat:
    post:
      operationId: sendChatMessage
      tags: [ai]
      summary: Send a message to the AI assistant (creates a conversation if none given)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: Assistant reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
...
components:
  schemas:
...
    ErrorMessage:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    ComplianceCheck:
      type: object
      required: [id, locationId, category, description, status, score, checkedAt]
      properties:
        id: { type: integer }
        locationId: { type: integer }
        category: { type: string }
        description: { type: string }
        status: { type: string, enum: [pass, warn, fail] }
        score: { type: integer }
        checkedAt: { type: string }
        dueDate: { type: ["string", "null"] }
    ComplianceCheckInput:
      type: object
      required: [locationId, category, description, status, score]
      properties:
        locationId: { type: integer }
        category: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        status: { type: string, enum: [pass, warn, fail] }
        score: { type: integer, minimum: 0, maximum: 100 }
        dueDate: { type: string }
    ComplianceCheckUpdate:
      type: object
      properties:
        category: { type: string }
        description: { type: string }
        status: { type: string, enum: [pass, warn, fail] }
        score: { type: integer, minimum: 0, maximum: 100 }
        dueDate: { type: string }
...
    Audit:
      type: object
      required: [id, locationId, auditor, score, conductedAt]
      properties:
        id: { type: integer }
        locationId: { type: integer }
        auditor: { type: string }
        score: { type: integer }
        notes: { type: ["string", "null"] }
        conductedAt: { type: string }
    AuditInput:
      type: object
      required: [locationId, auditor, score]
      properties:
        locationId: { type: integer }
        auditor: { type: string, minLength: 1 }
        score: { type: integer, minimum: 0, maximum: 100 }
        notes: { type: string }
        conductedAt: { type: string }
    Document:
      type: object
      required: [id, name, category, uploadedAt]
      properties:
        id: { type: integer }
        locationId: { type: ["integer", "null"] }
        taskId: { type: ["integer", "null"] }
        name: { type: string }
        category: { type: string }
        content: { type: ["string", "null"] }
        uploadedAt: { type: string }
    DocumentInput:
      type: object
      required: [name, category]
      properties:
        locationId: { type: integer }
        taskId: { type: integer }
        name: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        content: { type: string }
...
    ChatMessage:
      type: object
      required: [id, conversationId, role, content, createdAt]
      properties:
        id: { type: integer }
        conversationId: { type: integer }
        role: { type: string, enum: [user, assistant] }
        content: { type: string }
        createdAt: { type: string }
    ChatRequest:
      type: object
      required: [message]
      properties:
        conversationId: { type: integer }
        message: { type: string, minLength: 1 }
    ChatResponse:
      type: object
      required: [conversationId, reply]
      properties:
        conversationId: { type: integer }
        reply: { type: string }
...