# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FranchiseIntelligenceOS API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatRequest {
  conversationId?: number;
  /** @minLength 1 */
  message: string;
}